folder restructuring for express app
